+++ govtool/frontend/src/utils/openInNewTab.ts
@@ -7,6 +7,10 @@ export const openInNewTab = (url: string) => {
       ? `${import.meta.env.VITE_IPFS_GATEWAY}/${url.slice(7)}`
       : `https://${url}`;
 
+  // eslint-disable-next-line no-console
+  console.debug("Opening in new tab", fullUrl);
+
+  // Open the URL in a new tab
   const newWindow = window.open(fullUrl, "_blank", "noopener,noreferrer");
   if (newWindow) newWindow.opener = null;
 };